# docs/archive/Agentpay_snapshot_2026-03-16/backups/sensitive_backup_20260312_073532/openapi.yaml
openapi: 3.1.0
info:
  title: AgentPay API
  description: |
    **The Financial OS for AI Agents** — Universal payment gateway with AgentRank trust scoring,
    A2A escrow, multi-protocol support (x402, ACP, AP2, Solana Pay, Stripe), and KYA identity.

    ## Authentication
    All protected endpoints require an API key via:
    - `Authorization: Bearer <api_key>` header, or
    - `X-Api-Key: <api_key>` header

    ## Protocols Supported
    - **x402** — HTTP 402 paywall standard
    - **ACP** — Agent Communication Protocol
    - **AP2** — Agent Payment Protocol v2
    - **Solana Pay** — Solana USDC native
    - **Stripe** — Fiat card / bank payments

    Get your API key at [https://dashboard.agentpay.gg](https://dashboard.agentpay.gg)
  version: 1.0.0
  contact:
    name: AgentPay Support
    url: https://docs.agentpay.gg
    email: support@agentpay.gg
  license:
    name: MIT
    url: https://github.com/Rumblingb/Agentpay/blob/main/LICENSE
servers:
  - url: https://api.agentpay.gg
    description: Production
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/agentrank/{agentId}:
    get:
      tags: [AgentRank]
      summary: Get AgentRank score
      description: Returns the AgentRank trust score (0-1000) and breakdown for an agent
      operationId: getAgentRank
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: AgentRank data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRankResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/agentrank/leaderboard:
    get:
      tags: [AgentRank]
      summary: AgentRank leaderboard
      description: Returns top agents ranked by AgentRank score with pagination
      operationId: getLeaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: tier
          in: query
          schema:
            type: string
            enum: [S, A, B, C, D, F]
        - name: minScore
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 1000
      responses:
        '200':
          description: Leaderboard results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Format: 'Bearer <api_key>' or just '<api_key>'"
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            type: string
...
    AgentRankResponse:
      type: object
      properties:
        agentId:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 1000
        tier:
          type: string
          enum: [bronze, silver, gold, platinum, diamond]
        breakdown:
          type: object
          properties:
            paymentReliability:
              type: number
            serviceDelivery:
              type: number
            transactionVolume:
              type: number
            walletAge:
              type: number
            disputeRate:
              type: number
...
    LeaderboardResponse:
      type: object
      properties:
        success:
          type: boolean
        leaderboard:
          type: array
          items:
            type: object
            properties:
              rank:
                type: integer
              agentId:
                type: string
              score:
                type: integer
                minimum: 0
                maximum: 1000
              grade:
                type: string
              transactionVolume:
                type: integer
              paymentReliability:
                type: number
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean